+++ i18n/zh-CN/docusaurus-plugin-content-docs-api/current/docs/docs-api/more-information/changelog.md
@@ -68,6 +68,7 @@ ONLYOFFICE 文档 API 的更改列表。
 - 向 WOPI 主机页面的 *form* 元素添加了 [docs\_api\_config](../using-wopi/host-page.md#parameters) 参数。
 - 向转换请求添加了 [pdf](../additional-api/conversion-api/request.md#pdf) 字段。
 - 添加了 [events.onSubmit](../usage-api/config/events.md#onsubmit) 事件。
+- 添加了 [events.onSaveDocument](../usage-api/config/events.md#onsavedocument) 事件。
 - 将 *roles* 字段添加到 [editorConfig.customization.features](../usage-api/config/editor/customization/customization-standard-branding.md#features) 参数。
 - 发送时将 [shardkey](../get-started/configuration/shard-key.md) 参数添加到 URL 查询字符串请求到文档命令服务、文档转换服务或文档生成器服务。
 - 向 *Automation API* 添加了 [addContextMenuItem](../usage-api/automation-api/connector-class.md#addcontextmenuitem), [addToolbarMenuItem](../usage-api/automation-api/connector-class.md#addtoolbarmenuitem) 和 [updateContextMenuItem](../usage-api/automation-api/connector-class.md#updatecontextmenuitem) 方法。